as: Declare testing variables locally
